Extract shell control-plane DTOs
