+++ clients/apple/scripts/test-terminal-surface-controller.sh
@@ -12,6 +12,7 @@ mkdir -p "$MODULE_CACHE_DIR"
 CLANG_MODULE_CACHE_PATH="$MODULE_CACHE_DIR" swiftc \
     "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellValueTypes.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellSnapshots.swift" \
+    "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellControlPlaneDTOs.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellTreeMutations.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/Models/Shell/ShellStateMutations.swift" \
     "$REPO_ROOT/clients/apple/AlanNative/ShellModel.swift" \
